+++ react/RatingInline.tsx
@@ -7,6 +7,7 @@ import { useProduct } from 'vtex.product-context'
 import Stars from './components/Stars'
 import TotalReviewsByProductId from '../graphql/totalReviewsByProductId.graphql'
 import AverageRatingByProductId from '../graphql/averageRatingByProductId.graphql'
+import AppSettings from '../graphql/appSettings.graphql'
 
 interface TotalData {
   totalReviewsByProductId: number
@@ -16,22 +17,54 @@ interface AverageData {
   averageRatingByProductId: number
 }
 
+interface SettingsData {
+  appSettings: AppSettings
+}
+
+interface AppSettings {
+  allowAnonymousReviews: boolean
+  requireApproval: boolean
+  useLocation: boolean
+  defaultOpen: boolean
+  defaultOpenCount: number
+  showGraph: boolean
+  displaySummaryIfNone: boolean
+  displayInlineIfNone: boolean
+  displaySummaryTotalReviews: boolean
+  displaySummaryAddButton: boolean
+}
+
 interface State {
   total: number
   average: number
   hasTotal: boolean
   hasAverage: boolean
+  settings: AppSettings
 }
 
 type ReducerActions =
   | { type: 'SET_TOTAL'; args: { total: number } }
   | { type: 'SET_AVERAGE'; args: { average: number } }
+  | { type: 'SET_SETTINGS'; args: { settings: AppSettings } }
+
 
 const initialState = {
   total: 0,
   average: 0,
   hasTotal: false,
   hasAverage: false,
+  settings: {
+    defaultOpen: false,
+    defaultOpenCount: 0,
+    allowAnonymousReviews: false,
+    requireApproval: true,
+    useLocation: false,
+    showGraph: false,
+    displaySummaryIfNone: false,
+    displayInlineIfNone: false,
+    displaySummaryTotalReviews: true,
+    displaySummaryAddButton: false,
+  },
 }
 
 const reducer = (state: State, action: ReducerActions) => {
@@ -47,6 +80,11 @@ const reducer = (state: State, action: ReducerActions) => {
         ...state,
         average: action.args.average,
         hasAverage: true,
+      }
+      case 'SET_SETTINGS':
+      return {
+        ...state,
+        settings: action.args.settings,
       }
     default:
       return state
@@ -82,8 +120,8 @@ function RatingInline() {
           args: { total },
         })
       })
-
-    client
+      
+      client
       .query({
         query: AverageRatingByProductId,
         variables: {
@@ -97,12 +135,25 @@ function RatingInline() {
           args: { average },
         })
       })
-  }, [client, productId])
+
+      client
+      .query({
+        query: AppSettings,
+      })
+      .then((response: ApolloQueryResult<SettingsData>) => {
+      const settings = response.data.appSettings
+      dispatch({
+        type: 'SET_SETTINGS',
+        args: { settings },
+      })
+    })
+    }, [client, productId])
+    
 
   return (
     <div className={`${handles.inlineContainer} review-summary mw8 center`}>
       {!state.hasTotal || !state.hasAverage ? null : state.total ===
-        0 ? null : (
+        0 && !state.settings.displayInlineIfNone ? null : (
         <Fragment>
           <span className="t-heading-5 v-mid">
             <Stars rating={state.average} />