+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add setting to display stars in `product-rating-inline` block when the product has no reviews
+
 ## [2.11.8] - 2021-08-30
 
 ### Fixed